Remove DecodeBase58 duplication
Move base58 enc/dec functions out of base58.cpp, into a specific compilation unit.
This is necessary to remove code duplication in mcTestCall, and yet link correctly on all platforms.
